+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.9-dev+4.11.2014
 
+- Added the `scrollToLine` and `scrollToRow` methods to `Editor`.
+
 ## 0.3.8+4.11.2014
 
 - Added the `LanguageTools`, `AutoCompleter` and `Completion` classes to support